Get serverSpec test to pass
The server app functionality works (all the LiveServerIntegrationSpec tests pass
and the client works like it should). However since we are using fs read and write
file functions (which are asynchronous), they are making the serverSpec tests to fail.
We modified the serverSpec to include done() functions to make the tests wait for
handler.requestHandler to finish. This makes the tests pass but also adds duplicate failing
tests for some reason, and we don't know the reason at this time.